Revert "WIP: Rename PubspecReader to ConfigReader"
This reverts commit 25f5d88dd86f003e0e33f3ad9f06d21b14bb9d97.
