+++ packages/patrol_cli/lib/src/features/test/pubspec_reader.dart
