+++ packages/patrol_cli/lib/src/features/test/test_command.dart
@@ -9,9 +9,9 @@ import 'package:patrol_cli/src/features/run_commons/dart_defines_reader.dart';
 import 'package:patrol_cli/src/features/run_commons/device.dart';
 import 'package:patrol_cli/src/features/run_commons/test_finder.dart';
 import 'package:patrol_cli/src/features/test/android_test_backend.dart';
-import 'package:patrol_cli/src/features/test/config_reader.dart';
 import 'package:patrol_cli/src/features/test/ios_test_backend.dart';
 import 'package:patrol_cli/src/features/test/native_test_runner.dart';
+import 'package:patrol_cli/src/features/test/pubspec_reader.dart';
 
 import '../../common/staged_command.dart';
 import '../../common/tool_exit.dart';
